Added dynamic output paths and user feedback for saved files

+++ src/main/scala/models/Output/FlightsByMonth.scala
@@ -1,3 +1,3 @@
 package models.Output
 
-case class FlightsByMonth(month: Long, totalNumberOfFlights : Int)
+case class FlightsByMonth(month: Int, totalNumberOfFlights : Int)
